securedrop/requirements/securedrop-requirements.txt
beautifulsoup4==4.3.2
Flask-WTF==0.11
Flask==0.10.1
Flask-Babel==0.9
gnupg-securedrop==1.2.5-9-g6f9d63a-dirty
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pip-tools==0.3.5
psutil==2.2.0
py==1.4.26
pycrypto==2.6.1
pyotp==1.4.2
python-gettext==2.1
qrcode==5.1
redis==2.10.3
rq==0.4.6
scrypt==0.6.1
six==1.9.0
SQLAlchemy==0.9.8
Werkzeug==0.9.6
WTForms==2.0.2
